requirements.txt
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
bcrypt==5.0.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.5
distro==1.9.0
ecdsa==0.19.1
fastapi==0.135.1
greenlet==3.3.2
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
jiter==0.13.0
Mako==1.3.10
MarkupSafe==3.0.3
numpy==2.4.2
openai==2.24.0
passlib==1.7.4
pgvector==0.4.2
psycopg2-binary==2.9.11
pyasn1==0.6.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pypdf==6.7.5
python-dotenv==1.2.2
python-jose==3.5.0
PyYAML==6.0.3
regex==2026.2.28
requests==2.32.5
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.48
starlette==0.52.1
tiktoken==0.12.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.41.0
watchfiles==1.1.1
websockets==16.0




# SpeechBrain pour le darija
speechbrain==1.0.0
torch==2.5.1
torchaudio==2.5.1
transformers==4.46.3
sentencepiece==0.2.0